url externe dans meta données

+++ modules/traitementOeuvre.php
@@ -186,7 +186,7 @@ if (isset($_POST['idOeuvre'])) {
 		
 	 }
 
-//traitement contenu +
+	//TRAITEMENT CONTENU +
 	 
 	
 	
@@ -199,53 +199,65 @@ if (isset($_POST['idOeuvre'])) {
 	if (isset($_POST['typeDonnee'])) {
 		$idType = htmlentities($_POST['typeDonnee']);
 	}
-	if (isset($_POST['libelleDonnee'])) {
-		$libelleDonnee = htmlentities($_POST['libelleDonnee']);
-		//on enleve les espace pour le nom de fichier
-		$space = explode(" ", $libelleDonnee);
-		$nomFichier = '';
-		foreach ($space as $key => $value) {
-			$nomFichier .= $value;
+
+
+	//si le type est différent de 4 (lien externe) on effectue la suite  sinon on fait autrement
+	
+		
+		if (isset($_POST['libelleDonnee'])) {
+			$libelleDonnee = htmlentities($_POST['libelleDonnee']);
+			//on enleve les espace pour le nom de fichier
+			$space = explode(" ", $libelleDonnee);
+			$nomFichier = '';
+			foreach ($space as $key => $value) {
+				$nomFichier .= $value;
+			}
+		}
+
+	if (isset($idType) && $idType != 4 ) {
+
+		if (isset($_FILES['fichierDonnee']) && ($_FILES['fichierDonnee']['name'][0] != NULL)) {
+			$name = $_FILES['fichierDonnee']['name'][0];
+			
+			if (htmlentities(isset($_POST['MAX_FILE_SIZE'])) && $_POST['MAX_FILE_SIZE'] == '500000') {
+				$maxsize = (int)$_POST['MAX_FILE_SIZE'];
+				if ($_FILES['fichierDonnee']['error'][0] > 0) $erreur = "Erreur lors du transfert";
+				if ($_FILES['fichierDonnee']['size'][0] > $maxsize) {$erreur = "Le fichier est trop gros";}
+				
+				$extensions_valides = array( 'jpg' , 'jpeg' , 'gif' , 'png', 'mp3', 'mp4', 'wav', 'mpeg');
+				$extension_upload = strtolower(  substr(  strrchr($name, '.')  ,1)  );
+				// if ( in_array($extension_upload,$extensions_valides) ) echo "Extension correcte";
+				$cheminFichier = "../meta/oeuvre{$idOeuvre}/{$nomFichier}.{$extension_upload}";
+				//suppression des fichiers existants
+				$file = "../meta/oeuvre{$idOeuvre}/{$nomFichier}.{$extension_upload}";
+				if (file_exists($file)) {
+					unlink($file);
+				}
+				$resultat = move_uploaded_file($_FILES['fichierDonnee']['tmp_name'][0],$cheminFichier);
+				// if ($resultat) echo "Transfert réussi";
+				$nomFichier = $nomFichier.'.'.$extension_upload;
+			}
+			
+		}
+
+
+	}else{
+		if (isset($_POST['lien'])) {
+			$nomFichier = htmlentities($_POST['lien']);
 		}
 	}
 
-	if (isset($_FILES['fichierDonnee']) && ($_FILES['fichierDonnee']['name'][0] != NULL)) {
-		$name = $_FILES['fichierDonnee']['name'][0];
-		
-		if (htmlentities(isset($_POST['MAX_FILE_SIZE'])) && $_POST['MAX_FILE_SIZE'] == '500000') {
-			$maxsize = (int)$_POST['MAX_FILE_SIZE'];
-			if ($_FILES['fichierDonnee']['error'][0] > 0) $erreur = "Erreur lors du transfert";
-			if ($_FILES['fichierDonnee']['size'][0] > $maxsize) {$erreur = "Le fichier est trop gros";}
-			
-			$extensions_valides = array( 'jpg' , 'jpeg' , 'gif' , 'png', 'mp3', 'mp4', 'wav', 'mpeg');
-			$extension_upload = strtolower(  substr(  strrchr($name, '.')  ,1)  );
-			// if ( in_array($extension_upload,$extensions_valides) ) echo "Extension correcte";
-			$cheminFichier = "../meta/oeuvre{$idOeuvre}/{$nomFichier}.{$extension_upload}";
-			//suppression des fichiers existants
-			$file = "../meta/oeuvre{$idOeuvre}/{$nomFichier}.{$extension_upload}";
-			if (file_exists($file)) {
-				unlink($file);
-			}
-			$resultat = move_uploaded_file($_FILES['fichierDonnee']['tmp_name'][0],$cheminFichier);
-			// if ($resultat) echo "Transfert réussi";
-			$nomFichier = $nomFichier.'.'.$extension_upload;
-			//mise a jour de la base
-			$donnee = new DonneeEnrichie(['urlFichier'=>$nomFichier, 'libelleDonneeEnrichie'=>$libelleDonnee, 'idTypeDonneEnrichie'=>$idType, 'idOeuvre'=>$idOeuvre]);
-			$managerMeta = new DonneeEnrichieManager($bdd);
-			$managerMeta->addDonnee($donnee);
-			
-			//recuperation de l'id du dernier ajout en base pour l'affichage ajax
-			$lastId = $managerMeta->getLastDonnee();
-			echo $lastId;
-			// header('location: ../content/gestionPanel.php');
-			
-	 	}
-		
-	 }
-
+	//mise a jour de la base
+	$donnee = new DonneeEnrichie(['urlFichier'=>$nomFichier, 'libelleDonneeEnrichie'=>$libelleDonnee, 'idTypeDonneEnrichie'=>$idType, 'idOeuvre'=>$idOeuvre]);
+	$managerMeta = new DonneeEnrichieManager($bdd);
+	$managerMeta->addDonnee($donnee);
+	
+	//recuperation de l'id du dernier ajout en base pour l'affichage ajax
+	$lastId = $managerMeta->getLastDonnee();
+	echo $lastId;
+	// header('location: ../content/gestionPanel.php');	
 
 }
-
 
 //suppression contenu +
 if (isset($_GET['req'], $_GET['idOeuvre']) && $_GET['req'] == 'deleteMeta') {
